commit 5

+++ lib/screens/home_screen/widgets/journal_card.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_web_api_v2/helpers/weekday.dart';
 import 'package:flutter_web_api_v2/models/journal.dart';
+import 'package:flutter_web_api_v2/screens/commom/confirmation_dialog.dart';
 import 'package:flutter_web_api_v2/services/journal_service.dart';
 import 'package:uuid/uuid.dart';
 
@@ -113,14 +114,25 @@ class JournalCard extends StatelessWidget {
   void removeJournal(BuildContext context) {
     JournalService service = JournalService();
     if (journal != null) {
-      service.delete(journal!.id).then((value) {
-        if (value) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text("Removido com sucesso!"),
-            ),
-          );
-          refreshFunction();
+      showConfirmationDialog(
+        context,
+        content:
+            "Deseja realmente remover o diário de ${WeekDay(journal!.createdAt)}?",
+        affirmativeOption: "Remover",
+      ).then((value) {
+        if (value != null) {
+          if (value) {
+            service.delete(journal!.id).then((value) {
+              if (value) {
+                ScaffoldMessenger.of(context).showSnackBar(
+                  const SnackBar(
+                    content: Text("Removido com sucesso!"),
+                  ),
+                );
+                refreshFunction();
+              }
+            });
+          }
         }
       });
     }
